+++ scripts/restart.sh
@@ -1,5 +1,5 @@
 #!/bin/bash
-
-docker compose restart api-gateway
+docker compose down api-gateway
+docker compose up -d api-gateway
 sleep 2
 docker logs -f --since 2s kodecraft-gateway